% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chao2_sac_functions.R
\name{calculate_inventory_completeness_batch}
\alias{calculate_inventory_completeness_batch}
\title{Batch Calculate Inventory Completeness Metrics}
\usage{
calculate_inventory_completeness_batch(inputdir, outputdir, shapefiledir)
}
\arguments{
\item{inputdir}{Character. Path to directory containing site CSV files.}

\item{outputdir}{Character. Path to save results.}

\item{shapefiledir}{Character. Path to WKT files.}
}
\value{
A data frame with site IDs and completeness metrics.
}
\description{
Batch Calculate Inventory Completeness Metrics
}
