% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_year_occurrence.R
\name{calculate_mean_year_batch}
\alias{calculate_mean_year_batch}
\title{Calculate Mean Year for All Sites in a Directory}
\usage{
calculate_mean_year_batch(inputdir, outputdir, shapefiledir)
}
\arguments{
\item{inputdir}{Character. Path to directory containing site CSV files.}

\item{outputdir}{Character. Path to save results.}

\item{shapefiledir}{Character. Path to WKT files.}
}
\value{
A data frame with site IDs and mean years.
}
\description{
Calculate Mean Year for All Sites in a Directory
}
