% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialized_indicators.R
\name{calculate_specialized_batch}
\alias{calculate_specialized_batch}
\title{Batch Calculate Specialized Indicators}
\usage{
calculate_specialized_batch(
  inputdir,
  outputdir,
  shapefiledir,
  tree_path = NULL,
  eicat_data = NULL
)
}
\arguments{
\item{inputdir}{Character. Path to directory containing site CSV files.}

\item{outputdir}{Character. Path to save results.}

\item{shapefiledir}{Character. Path to WKT files.}

\item{tree_path}{Character. Path to phylogenetic tree file.}

\item{eicat_data}{Data frame. EICAT assessment data.}
}
\value{
A data frame with specialized indicators.
}
\description{
Batch Calculate Specialized Indicators
}
