% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gbif_credentials}
\alias{check_gbif_credentials}
\title{Check for GBIF Credentials}
\usage{
check_gbif_credentials()
}
\value{
Logical. TRUE if all credentials are set, stops otherwise.
}
\description{
Verifies that the required GBIF credentials (user, pwd, email) are set
in the R options.
}
