% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_violin_plot.R
\name{create_violin_plot}
\alias{create_violin_plot}
\title{Create Violin Plot of Site Occurrences}
\usage{
create_violin_plot(nested_list)
}
\arguments{
\item{nested_list}{A nested list of countries and sites with occurrence values.}
}
\value{
A ggplot object.
}
\description{
Create Violin Plot of Site Occurrences
}
