% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_year_occurrence.R
\name{filter_by_shapefile}
\alias{filter_by_shapefile}
\title{Filter cube data by shapefile boundary}
\usage{
filter_by_shapefile(cube, shapefile_path)
}
\arguments{
\item{cube}{Data frame. The GBIF occurrence cube.}

\item{shapefile_path}{Character. Path to WKT file.}
}
\value{
Filtered data frame.
}
\description{
Filter cube data by shapefile boundary
}
