% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{get_gbif_predicates}
\alias{get_gbif_predicates}
\title{Generate GBIF Predicates from WKT and Filters}
\usage{
get_gbif_predicates(wkt, filters)
}
\arguments{
\item{wkt}{A WKT string defining the geometry.}

\item{filters}{A list of filters to apply (e.g., hasGeospatialIssue, basisOfRecord).}
}
\value{
A GBIF predicate object.
}
\description{
Generate GBIF Predicates from WKT and Filters
}
