% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_orchestrator.R
\name{get_site_list}
\alias{get_site_list}
\title{Get Site List from Data Directory}
\usage{
get_site_list(data_dir)
}
\arguments{
\item{data_dir}{Character. Path to data directory.}
}
\value{
A list of site information.
}
\description{
Get Site List from Data Directory
}
