% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_manager.R
\name{load_download_key}
\alias{load_download_key}
\title{Load Download Key from Disk}
\usage{
load_download_key(path = ".gbif_download_key")
}
\arguments{
\item{path}{File path to the saved key.}
}
\value{
The key as character, or NULL if not found.
}
\description{
Load Download Key from Disk
}
