% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_global_results.R
\name{perform_data_gap_analysis}
\alias{perform_data_gap_analysis}
\title{Perform Data Gap Analysis}
\usage{
perform_data_gap_analysis(summary_data)
}
\arguments{
\item{summary_data}{Data frame from generate_global_summary().}
}
\value{
A list with analysis results.
}
\description{
Perform Data Gap Analysis
}
