% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_manager.R
\name{save_download_key}
\alias{save_download_key}
\title{Save Download Key to Disk}
\usage{
save_download_key(key, path = ".gbif_download_key")
}
\arguments{
\item{key}{The GBIF download key.}

\item{path}{File path to save the key.}
}
\description{
Save Download Key to Disk
}
