% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{save_ramsar_plot}
\alias{save_ramsar_plot}
\title{Plot and Save a Ramsar Biodiversity Indicator}
\usage{
save_ramsar_plot(
  indicator_obj,
  filename,
  device = "png",
  width = 4000,
  height = 4000,
  units = "px",
  ...
)
}
\arguments{
\item{indicator_obj}{A b3gbi_indicator object.}

\item{filename}{The path to save the plot (including extension).}

\item{device}{The device to use (default: "png").}

\item{width}{Width of the plot (default: 4000).}

\item{height}{Height of the plot (default: 4000).}

\item{units}{Units for width and height (default: "px").}

\item{...}{Additional arguments passed to b3gbi::plot.}
}
\value{
The ggplot object (invisibly).
}
\description{
Plot and Save a Ramsar Biodiversity Indicator
}
